+++ app/src/main/java/campus/tech/kakao/map/MainActivity.kt
@@ -20,39 +20,17 @@ class MainActivity : AppCompatActivity() {
             mapItemList.add(MapItem("카페" + i, "복현동" + i, "카페"))
         }
 
+        val selectList = findViewById<RecyclerView>(R.id.selectList)
+        val selectItemList = mutableListOf<SelectItem>()
+        for (i in 0..20) {
+            selectItemList.add(SelectItem("카페" + i))
+        }
+
         mapList.adapter = MapListAdapter(mapItemList, LayoutInflater.from(this))
         mapList.layoutManager = LinearLayoutManager(this, LinearLayoutManager.VERTICAL, false)
-    }
-}
 
-class MapListAdapter(
-    val mapItemList: MutableList<MapItem>, val layoutInflater: LayoutInflater
-) : RecyclerView.Adapter<MapListAdapter.ViewHolder>() {
-    inner class ViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
-        val name: TextView
-        val address: TextView
-        val category: TextView
-
-        init {
-            name = itemView.findViewById<TextView>(R.id.name)
-            address = itemView.findViewById<TextView>(R.id.address)
-            category = itemView.findViewById<TextView>(R.id.category)
-        }
-    }
-
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
-        val view = layoutInflater.inflate(R.layout.map_item, parent, false)
-        return ViewHolder(view)
-    }
-
-    override fun onBindViewHolder(holder: ViewHolder, position: Int) {
-        holder.name.text = mapItemList.get(position).name
-        holder.address.text = mapItemList.get(position).address
-        holder.category.text = mapItemList.get(position).category
-    }
-
-    override fun getItemCount(): Int {
-        return mapItemList.size
+        selectList.adapter = SelectListAdapter(selectItemList, LayoutInflater.from(this))
+        selectList.layoutManager = LinearLayoutManager(this, LinearLayoutManager.HORIZONTAL, false)
     }
 }
