+++ app/src/main/res/layout/select_item.xml
@@ -5,6 +5,7 @@
     android:layout_height="30dp">
 
     <TextView
+        android:id="@+id/name"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:text="카페1"
